ensure all drawing object always has a drawing order > 0 set

+++ OpenFTTH.Schematic.API/Model/DiagramLayout/DiagramObject.cs
@@ -13,7 +13,7 @@ namespace OpenFTTH.Schematic.API.Model.DiagramLayout
         private readonly Diagram _diagram;
         public Diagram Diagram => _diagram;
 
-        public UInt16 DrawingOrder { get; set; }
+        public int DrawingOrder { get; set; }
 
         // Geometry
         public Geometry Geometry { get; init; }
