+++ OpenFTTH.Schematic.Business/InternalDiagramObjects/Lines/LineBlockConnection.cs
@@ -11,7 +11,7 @@ namespace OpenFTTH.Schematic.Business.Lines
 
         public string Label { get; set; }
         public string Style { get; set; }
-        public ushort DrawingOrder { get; set; }
+        public int DrawingOrder { get; set; }
 
         internal IEnumerable<DiagramObject> CreateDiagramObjects(Diagram diagram)
         {
